+++ package.json
@@ -26,6 +26,7 @@
 		"express": "^4.17.1",
 		"helmet": "^3.21.2",
 		"knex": "^0.20.8",
+		"react-router-dom": "^5.1.2",
 		"sqlite3": "^4.1.1"
 	},
 	"devDependencies": {
