+++ projects/src/state/actionCreators.js
@@ -23,6 +23,25 @@ export const getProjects = () => dispatch => {
 };
 
 // Step 7: Design action creator functions
+export const getProject = id => dispatch => {
+	axios
+		.get("/api/projects/" + id)
+		.then(response => {
+			dispatch({
+				type: types.GET_PROJECT,
+				payload: {
+					project: response.data
+				}
+			});
+		})
+		.catch(error => {
+			console.log(error);
+
+			console.log("error!!!");
+		});
+};
+
+// Step 7: Design action creator functions
 export const deleteProject = id => dispatch => {
 	axios
 		.delete("/api/projects/" + id)
