+++ projects/src/state/reducers.js
@@ -2,7 +2,10 @@ import * as types from "./actionTypes";
 
 // STEP 1: Decide state slices
 const initialState = {
-	projects: []
+	projects: [],
+	project: {
+		actions: []
+	}
 };
 
 // STEP 3: create reducers
@@ -17,5 +20,15 @@ export function projectsReducer(projects = initialState.projects, action) {
 			);
 		default:
 			return projects;
+	}
+}
+
+// STEP 3: create reducers
+export function projectReducer(project = initialState.project, action) {
+	switch (action.type) {
+		case types.GET_PROJECT:
+			return action.payload.project;
+		default:
+			return project;
 	}
 }
